+++ src/lxc/utils.h
@@ -173,9 +173,9 @@ extern int wait_for_pid(pid_t pid);
 extern int lxc_wait_for_pid_status(pid_t pid);
 
 /* send and receive buffers completely */
-extern int lxc_write_nointr(int fd, const void* buf, size_t count);
-extern int lxc_read_nointr(int fd, void* buf, size_t count);
-extern int lxc_read_nointr_expect(int fd, void* buf, size_t count, const void* expected_buf);
+extern ssize_t lxc_write_nointr(int fd, const void* buf, size_t count);
+extern ssize_t lxc_read_nointr(int fd, void* buf, size_t count);
+extern ssize_t lxc_read_nointr_expect(int fd, void* buf, size_t count, const void* expected_buf);
 #if HAVE_LIBGNUTLS
 #define SHA_DIGEST_LENGTH 20
 extern int sha1sum_file(char *fnam, unsigned char *md_value);
